+++ src/main/scala/ordogkatlan/ops/distribution/model/FulfilledWish.scala
@@ -9,7 +9,7 @@ case class FulfilledWish(
   visitorId: UUID,              //a látogató azonosítója
   wishId: UUID,                 //a kívánság azonosítója
   wonSeats: Int,                //megkapott helyek száma
-  credits: Float,               //a megkapott helyek összesített ára
+  credits: Double,              //a megkapott helyek összesített ára
   play: TicketablePlay,         //a kapcsolódó előadás
   priority: Int                 //a látogató által beállított prioritás, melyen a kívánság teljesült
 )
@@ -17,7 +17,7 @@ case class FulfilledWish(
 object FulfilledWish {
 
   //a teljesült kívánság felépítése a kívánságból, a kapcsolódó előadásból és a látogató azonosítójából
-  def from(wish: CalculableWish, play: TicketablePlay, applicantId: UUID):FulfilledWish = FulfilledWish(
+  def from(wish: CalculableWish, play: TicketablePlay, applicantId: UUID): FulfilledWish = FulfilledWish(
     visitorId = applicantId,
     wishId = wish.wishId,
     wonSeats = wish.wonSeats,
